+++ tests/pkgtttest/mkprojs.py
@@ -3,12 +3,12 @@
 from os import makedirs
 from os.path import join
 from os.path import exists
+from os.path import isabs
 from subprocess import run
 from logging import debug
 from collections import namedtuple
 
 from timetracker.consts import DIRTRK
-from timetracker.cmd.init import run_init
 
 RELCSVS = [
     "filename.csv",
@@ -56,22 +56,23 @@ def _get_expdirs(tmphome, project, dirgit, trksubdir):
         dirgit=join(dirproj, '.git') if dirgit else None,
         dirtrk=join(dirproj, trksubdir),
         dirdoc=join(dirproj, 'doc'))
-    prt_expdirs(ntexpdirs)
+    prt_expdirs(ntexpdirs, "tests/pkgtttest/mkprojs:mk_projdirs")
     return ntexpdirs
 
-def prt_expdirs(ntexpdirs):
+def prt_expdirs(ntexpdirs, name=""):
     """Print the expected directories and files and if they exist"""
     for key, expdir in ntexpdirs._asdict().items():
-        debug(f'exists({int(exists(expdir)) if expdir is not None else "."}) '
+        debug(f'{name} '
+              f'exists({int(exists(expdir)) if expdir is not None and isabs(expdir) else "."}) '
               f'{key:14} {expdir}')
 
-def mk_projdirs_wcfgs(tmp_home, project, trksubdir='.timetracker'):
-    """Make sub-directories & cfgs in a temporary directory for use in tests"""
-    ##dirproj = mk_projdirs(tmp_home, project)
-    ntexpdirs = mk_projdirs(tmp_home, project)
-    fname_cfgproj = join(ntexpdirs.dirproj, trksubdir, 'config')
-    cfg = run_init(fname_cfgproj, '.', project, dirhome=tmp_home)
-    return ntexpdirs.dirproj, cfg.cfg_loc, cfg.cfg_glb
+##def mk_projdirs_wcfgs(tmp_home, project, trksubdir='.timetracker'):
+##    """Make sub-directories & cfgs in a temporary directory for use in tests"""
+##    ##dirproj = mk_projdirs(tmp_home, project)
+##    ntexpdirs = mk_projdirs(tmp_home, project)
+##    fname_cfgproj = join(ntexpdirs.dirproj, trksubdir, 'config')
+##    cfg = run_init(fname_cfgproj, '.', project, dirhome=tmp_home)
+##    return ntexpdirs.dirproj, cfg.cfg_loc, cfg.cfg_glb
 
 def findhome(home):
     """Do a find on the given homedir and print using debug logging"""