+++ timetracker/cmd/init.py
@@ -18,30 +18,39 @@ def cli_run_init(fnamecfg, args):
             fnamecfg,
             args.csvdir,
             project=args.project,
-            fcfg_global=args.file)
+            trk_dir=args.trk_dir,
+            fcfg_global=args.global_config_file)
     else:
         run_reinit(
             fnamecfg,
             args.csvdir,
             project=args.project,
-            fcfg_global=args.file)
+            trk_dir=args.trk_dir,
+            fcfg_global=args.global_config_file)
 
-def run_init(fnamecfg, dircsv, project, **kwargs):
+def run_init(fnamecfg, dircsv=None, project=None, **kwargs):
     """Initialize timetracking on a project"""
-    cfg = Cfg(fnamecfg, kwargs.get('fcfg_global'), kwargs.get('dirhome'))
+    cfg = Cfg(fnamecfg)  #, kwargs.get('fcfg_global'), kwargs.get('dirhome'))
     # Initialize the local configuration file for a timetracking project
     cfg_loc = cfg.cfg_loc
     debug(yellow('RUNNING COMMAND INIT'))
     debug(f'INIT: fnamecfg:    {cfg_loc.filename}')
     debug(f'INIT: project:     {project}')
     debug(f'INIT: dircsv({dircsv})')
-    _chk_global_cfg(kwargs.get('fcfg_global'))
+    fcfg_global = kwargs.get('fcfg_global')
+    _chk_global_cfg(fcfg_global)
     if cfg_loc.exists:
         print(str_tostart())
         sys_exit(0)
     # WRITE A LOCAL PROJECT CONFIG FILE: ./.timetracker/config
-    cfg.init(project, dircsv)
-    debug(cfg.cfg_loc.get_desc("new"))
+    if project is None:
+        project = cfg.cfg_loc.get_project_from_filename()
+    if dircsv is None:
+        dircsv = '.'
+    debug('CMD INIT <<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<')
+    cfg.init(project, dircsv, fcfg_global, kwargs.get('dirhome'))
+    debug('CMD INIT >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>')
+    debug(cfg.cfg_loc.get_desc("ran_init"))
     return cfg
 
 def _chk_global_cfg(fcfg_global):
@@ -51,8 +60,8 @@ def _chk_global_cfg(fcfg_global):
 
 def run_reinit(fnamecfg, dircsv, project, **kwargs):
     """Reinitialize timetracking project"""
-    cfg = Cfg(fnamecfg, kwargs.get('fcfg_global'), kwargs.get('dirhome'))
-    cfg.reinit(project, dircsv=dircsv)
+    cfg = Cfg(fnamecfg)
+    cfg.reinit(project,dircsv, kwargs.get('fcfg_global'), kwargs.get('dirhome'))
 
 
 # Copyright (C) 2025-present, DV Klopfenstein, PhD. All rights reserved.